+++ app/assets/javascripts/components/projects/ProjectCard.jsx
@@ -5,50 +5,57 @@ export default class ProjectCard extends React.Component {
     const txt = document.createElement('textarea');
     txt.innerHTML = html;
     return txt.value;
-  };
+  }
 
   renderTag(project) {
-    if (!project.get('tag_name')) { return; }
+    if (!project.get('tag_name')) {
+      return;
+    }
 
     const style = {
       backgroundColor: project.get('tag_bg_color'),
-      color: project.get('tag_fore_color')
+      color: project.get('tag_fore_color'),
     };
 
     return (
       <small className="card-tag" style={style}>
-        { project.get('tag_name') }
+        {project.get('tag_name')}
       </small>
     );
   }
 
-  projectName(){
-    return this.decodeHTML(this.props.project.get('name')); 
+  projectName() {
+    return this.decodeHTML(this.props.project.get('name'));
   }
 
   panelHeading() {
     const { project, joined } = this.props;
 
     if (joined) {
-      return(
+      return (
         <div className="panel-heading card-heading">
           <div className="project-card-header-container">
-            <a href={ project.get('path_to').project } className="card-title project-title">
-              { this.projectName() }              
+            <a
+              href={project.get('path_to').project}
+              className="card-title project-title"
+            >
+              {this.projectName()}
             </a>
-            { this.renderTag(project) }
+            {this.renderTag(project)}
           </div>
           <div className="icons pull-right">
-            <a href={ project.get('path_to').projectReports }
+            <a
+              href={project.get('path_to').projectReports}
               className="unstyled-link"
               data-toggle="tooltip"
               data-placement="top"
-              data-title={ I18n.t('reports') }
+              data-title={I18n.t('reports')}
             >
               <i className="mi md-20 heading-icon">insert_chart</i>
             </a>
 
-            <a href={ project.get('path_to').projectUsers }
+            <a
+              href={project.get('path_to').projectUsers}
               className="unstyled-link"
               data-toggle="tooltip"
               data-placement="top"
@@ -57,7 +64,8 @@ export default class ProjectCard extends React.Component {
             </a>
 
             <span className="dropdown">
-              <a className="unstyled-link"
+              <a
+                className="unstyled-link"
                 type="button"
                 data-toggle="dropdown"
                 aria-haspopup="true"
@@ -66,26 +74,33 @@ export default class ProjectCard extends React.Component {
                 <i className="mi md-20 heading-icon">settings</i>
               </a>
 
-              <ul className="dropdown-menu dropdown-menu-right" aria-labelledby="dLabel">
-                <li className="dropdown-header">
-                  { I18n.t('options') }
-                </li>
+              <ul
+                className="dropdown-menu dropdown-menu-right"
+                aria-labelledby="dLabel"
+              >
+                <li className="dropdown-header">{I18n.t('options')}</li>
 
                 <li>
-                  <a href={ project.get('path_to').projectSettings }
+                  <a
+                    href={project.get('path_to').projectSettings}
                     data-toggle="tooltip"
                     data-placement="top"
-                    data-title={ I18n.t('settings') }
+                    data-title={I18n.t('settings')}
                   >
-                    { I18n.t('settings') }
+                    {I18n.t('settings')}
                   </a>
                 </li>
 
                 <li className="divider"></li>
 
                 <li>
-                  <a href={`${project.get('path_to').projectUnjoin}${this.props.user.id}`} data-method="delete">
-                    { I18n.t('projects.unjoin') }
+                  <a
+                    href={`${project.get('path_to').projectUnjoin}${
+                      this.props.user.id
+                    }`}
+                    data-method="delete"
+                  >
+                    {I18n.t('projects.unjoin')}
                   </a>
                 </li>
               </ul>
@@ -95,11 +110,9 @@ export default class ProjectCard extends React.Component {
       );
     }
 
-    return(
+    return (
       <div className="panel-heading card-heading">
-        <span className="card-title">
-          { this.projectName() }
-        </span>
+        <span className="card-title">{this.projectName()}</span>
       </div>
     );
   }
@@ -108,28 +121,30 @@ export default class ProjectCard extends React.Component {
     const { project, joined } = this.props;
 
     if (joined) {
-      return(
+      return (
         <div className="panel-body">
           <div className="col-md-12 members">
-            <ul className="member-list">
-              { this.renderUsersAvatar() }
-            </ul>
+            <ul className="member-list">{this.renderUsersAvatar()}</ul>
           </div>
         </div>
       );
     }
 
     if (!joined && project.get('archived_at')) {
-      return(
+      return (
         <div className="panel-body">
-          <span className="col-md-12 text-center">{ I18n.t('projects.unable_to_join') }</span>
+          <span className="col-md-12 text-center">
+            {I18n.t('projects.unable_to_join')}
+          </span>
         </div>
       );
     }
 
-    return(
+    return (
       <div className="panel-body">
-        <span className="col-md-12 text-center">{ I18n.t('projects.to_view_more_join') }</span>
+        <span className="col-md-12 text-center">
+          {I18n.t('projects.to_view_more_join')}
+        </span>
       </div>
     );
   }
@@ -138,9 +153,11 @@ export default class ProjectCard extends React.Component {
     const { project, joined } = this.props;
 
     if (joined) {
-      return project.get('users_avatar').map((avatar_url) =>
-        <li key={ avatar_url } className="member"><img src={ avatar_url } alt="User avatar" className="identicon" /></li>
-      );
+      return project.get('users_avatar').map(avatar_url => (
+        <li key={avatar_url} className="member">
+          <img src={avatar_url} alt="User avatar" className="identicon" />
+        </li>
+      ));
     }
   }
 
@@ -148,23 +165,41 @@ export default class ProjectCard extends React.Component {
     const { project, joined } = this.props;
 
     if (project.get('archived_at')) {
-      return(<span className="card-footer panel-footer">{I18n.t('archived_at')} {project.get('archived_at')}</span>);
+      return (
+        <span className="card-footer panel-footer">
+          {I18n.t('archived_at')} {project.get('archived_at')}
+        </span>
+      );
     }
 
     if (joined) {
-      return (<a href={ project.get('path_to').project } className="card-footer panel-footer">{ I18n.t('projects.select') }</a>);
+      return (
+        <a
+          href={project.get('path_to').project}
+          className="card-footer panel-footer"
+        >
+          {I18n.t('projects.select')}
+        </a>
+      );
     }
 
-    return(<a href={ project.get('path_to').projectJoin } className="card-footer panel-footer">{ I18n.t('projects.join') }</a>);
+    return (
+      <a
+        href={project.get('path_to').projectJoin}
+        className="card-footer panel-footer"
+      >
+        {I18n.t('projects.join')}
+      </a>
+    );
   }
 
   render() {
     return (
       <div className="col-md-6">
         <div className="panel panel-default card project-item animated scale-in">
-          { this.panelHeading() }
-          { this.panelBody() }
-          { this.cardLink() }
+          {this.panelHeading()}
+          {this.panelBody()}
+          {this.cardLink()}
         </div>
       </div>
     );