+++ app/assets/javascripts/components/projects/SideBar/SideBarButton.jsx
@@ -3,20 +3,12 @@ import SideBarButtonInfo from './SideBarButtonInfo';
 import PropTypes from 'prop-types';
 import classname from 'classnames';
 
-const SideBarButton = ({
-  children,
-  description,
-  onClick,
-  isVisible
-}) => {
+const SideBarButton = ({ children, description, onClick, isVisible }) => {
   const [showInfo, setShowInfo] = useState(false);
 
-  const classes = classname(
-    'SideBar__link',
-    {
-      'SideBar__link--is-visible': isVisible
-    }
-  );
+  const classes = classname('SideBar__link', {
+    'SideBar__link--is-visible': isVisible,
+  });
 
   return (
     <li
@@ -26,19 +18,19 @@ const SideBarButton = ({
       onClick={onClick}
       data-id="side-bar-button"
     >
-      {
-        showInfo && <SideBarButtonInfo data-id="button-info" description={description} />
-      }
-      { children }
+      {showInfo && (
+        <SideBarButtonInfo data-id="button-info" description={description} />
+      )}
+      {children}
     </li>
-  )
-}
+  );
+};
 
 SideBarButton.propTypes = {
   children: PropTypes.node.isRequired,
   description: PropTypes.string.isRequired,
   onClick: PropTypes.func.isRequired,
-  isVisible: PropTypes.bool.isRequired
-}
+  isVisible: PropTypes.bool.isRequired,
+};
 
 export default SideBarButton;