+++ app/assets/javascripts/components/story/select_user/SelectUser.jsx
@@ -5,36 +5,28 @@ const SelectUser = ({ selectedUserId, onEdit, disabled, users }) => (
   <select
     value={selectedUserId || ''}
     className="form-control input-sm"
-    onChange={(event) => onEdit(event.target.value)}
+    onChange={event => onEdit(event.target.value)}
     disabled={disabled}
   >
-    <option value=''>
-      ----
-    </option>
-    {
-      users.map((user) => (
-        <option
-          value={user.id}
-          key={user.id}
-        >
-          {user.name}
-        </option>
-      ))
-    }
+    <option value="">----</option>
+    {users.map(user => (
+      <option value={user.id} key={user.id}>
+        {user.name}
+      </option>
+    ))}
   </select>
 );
 
 SelectUser.propTypes = {
-  selectedUserId: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.number
-  ]),
+  selectedUserId: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
   onEdit: PropTypes.func.isRequired,
-  users: PropTypes.arrayOf(PropTypes.shape({
-    id: PropTypes.number.isRequired,
-    name: PropTypes.string.isRequired
-  })),
-  disabled: PropTypes.bool.isRequired
+  users: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.number.isRequired,
+      name: PropTypes.string.isRequired,
+    })
+  ),
+  disabled: PropTypes.bool.isRequired,
 };
 
 export default SelectUser;