+++ app/assets/javascripts/controllers/tag_groups/TagGroupsController.jsx
@@ -3,15 +3,18 @@ import ReactDOM from 'react-dom';
 import ColorPicker from 'components/tag_groups/ColorPicker';
 
 export default () => {
-  const bgColor = $('[name=bg-color]').data('current-color') || "#F17013";
+  const bgColor = $('[name=bg-color]').data('current-color') || '#F17013';
 
-  ReactDOM.render(<ColorPicker color={ bgColor }/>, document.getElementById('tag_group_color'));
+  ReactDOM.render(
+    <ColorPicker color={bgColor} />,
+    document.getElementById('tag_group_color')
+  );
 
-  $(document).ajaxError(function( event, jqxhr, settings, thrownError ) {
-    $("#new_tag_group").renderFormErrors( $.parseJSON(jqxhr.responseText) );
+  $(document).ajaxError(function (event, jqxhr, settings, thrownError) {
+    $('#new_tag_group').renderFormErrors($.parseJSON(jqxhr.responseText));
   });
 
-  $.fn.modalSuccess = function(){
+  $.fn.modalSuccess = function () {
     // close modal
     this.modal('hide');
 
@@ -23,23 +26,25 @@ export default () => {
     this.clearPreviousErrors();
   };
 
-  $.fn.renderFormErrors = function(errors){
+  $.fn.renderFormErrors = function (errors) {
     let $form = this,
-        model = $form.data('model');
+      model = $form.data('model');
 
     // show error messages in input form-group help-block
-    $.each(errors, function(field, messages){
+    $.each(errors, function (field, messages) {
       let $input = $('input[name="' + model + '[' + field + ']"]');
       $input
         .closest('.form-group')
-        .addClass('has-error').find('.help-block').html( messages.join(' & ') );
+        .addClass('has-error')
+        .find('.help-block')
+        .html(messages.join(' & '));
     });
   };
 
-  $.fn.clearPreviousErrors = function(){
-    $('.form-group.has-error', this).each(function(){
+  $.fn.clearPreviousErrors = function () {
+    $('.form-group.has-error', this).each(function () {
       $('.help-block', $(this)).html('');
       $(this).removeClass('has-error');
     });
-  }
-}
+  };
+};