+++ app/assets/javascripts/libs/precompile_ejs.js
@@ -1,16 +1,16 @@
-import ejs from "ejs";
-import fs from "fs";
+import ejs from 'ejs';
+import fs from 'fs';
 
 export default function precompileEjs() {
   return {
-    name: "precompile-ejs",
-    enforce: "pre",
+    name: 'precompile-ejs',
+    enforce: 'pre',
     load: function (id) {
-      if (!id.endsWith(".ejs")) {
+      if (!id.endsWith('.ejs')) {
         return null;
       }
 
-      const code = fs.readFileSync(id).toString("utf-8");
+      const code = fs.readFileSync(id).toString('utf-8');
       const template = ejs.compile(code, {
         filename: id,
         client: true,
@@ -18,14 +18,14 @@ export default function precompileEjs() {
         rmWhitespace: true,
       });
       const transpiled =
-        "/** eslint-disable */\nexport default " + template.toString() + ";";
+        '/** eslint-disable */\nexport default ' + template.toString() + ';';
 
       return {
         code: transpiled,
       };
     },
     transform: function (src, id) {
-      if (!id.endsWith(".ejs")) {
+      if (!id.endsWith('.ejs')) {
         return null;
       }