+++ app/assets/javascripts/mixins/tab_notification.js
@@ -1,20 +1,22 @@
-
 const TabNotification = {
-  changeTitle: function(title, documentStatus) {
+  changeTitle: function (title, documentStatus) {
     var notificationTitle = ' \u2733 ' + title;
 
     if (documentStatus) {
-      this.handleChangeTitle(notificationTitle, $('body').data('notificationIconUrl'));
+      this.handleChangeTitle(
+        notificationTitle,
+        $('body').data('notificationIconUrl')
+      );
     } else {
       this.handleChangeTitle(title, $('body').data('faviconIconUrl'));
     }
   },
 
-  handleChangeTitle: function(title, href) {
+  handleChangeTitle: function (title, href) {
     var favicon = $('link[rel="shortcut icon"]');
     favicon.prop('href', href);
     document.title = title;
-  }
+  },
 };
 
 export default TabNotification;