+++ app/operations/project_operations/create.rb
@@ -2,15 +2,17 @@ module ProjectOperations
   class Create
     include Operation
 
-    def initialize(project:, current_user:)
+    def initialize(project:, current_user:, current_team:)
       @project = project
       @current_user = current_user
+      @current_team = current_team
     end
 
     def call
       ActiveRecord::Base.transaction do
         yield save_project
         yield create_activity
+        yield create_ownership
 
         Success(project)
       end
@@ -20,7 +22,7 @@ module ProjectOperations
 
     private
 
-    attr_reader :project, :current_user
+    attr_reader :project, :current_user, :current_team
 
     def save_project
       if project.save
@@ -36,6 +38,10 @@ module ProjectOperations
         current_user: current_user,
         action: 'create'
       )
+    end
+
+    def create_ownership
+      Success(current_team.ownerships.create!(project: project, is_owner: true))
     end
   end
 end