+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [3.0.0] 2023-10-10
+
+- Remove "Story attachments" feature
+- Implement real-time updates in the beta board
+- Refactor some components to be functional components
+
 ## [2.10.0] 2023-06-08
 
 - Refactor Note Operations to use dry-monads
@@ -653,7 +659,6 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
-[unreleased]: https://github.com/Codeminer42/cm42-central/compare/v2.8.0...HEAD
 [1.0.0]: https://github.com/Codeminer42/cm42-central/tree/v1.0.0
 [1.1.0]: https://github.com/Codeminer42/cm42-central/tree/v1.1.0
 [1.1.1]: https://github.com/Codeminer42/cm42-central/tree/v1.1.1
@@ -700,3 +705,4 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 [2.8.0]: https://github.com/Codeminer42/cm42-central/tree/v2.8.0
 [2.9.0]: https://github.com/Codeminer42/cm42-central/tree/v2.9.0
 [2.10.0]: https://github.com/Codeminer42/cm42-central/tree/v2.10.0
+[3.0.0]: https://github.com/Codeminer42/cm42-central/tree/v3.0.0